replace ExternalLinkDispay with ExternalLinkComponent
