+++ app/views/works/_show_info.html.erb
@@ -77,7 +77,7 @@
       <td>
         <ul>
           <% view.related_urls_filtered.each do |url| %>
-            <li class="attribute"><%= ExternalLinkDisplay.new(url).display %></li>
+            <li class="attribute"><%= render ExternalLinkComponent.new(url) %></li>
           <% end %>
         </ul>
       </td>
